+++ Sources/PlugIns/Graphic/Commands/RemoveGraphic.swift
@@ -1,8 +1,15 @@
 //
-//  File.swift
+//  RemoveGraphic.swift
 //  
 //
 //  Created by rrbox on 2023/08/20.
 //
 
-import Foundation
+import ECS
+
+class RemoveGraphic: EntityCommand {
+    override func runCommand(in world: World) {
+        world.removeGraphic(fromEntity: self.entity)
+    }
+}
+
